// src/enums/gameImages.ts
export enum GameImages {
   armor = 'armor',
   card = 'card',
   dead =  'dead',
   deathknight = 'deathknight',
   firedrake = 'firedrake',
   goldendragon = 'goldendragon',
   healingpotion = 'healingpotion',
   kobold = 'kobold',
   orge = 'ogre',
   paladin = 'paladin',
   playercard = 'playercard',
   restartbutton = 'restartbutton',
   shield = 'shield',
   troll = 'troll'
}